fix build problems
